new module and short schematic
racket module will handle all the rackets. They have to be configured in start-up and must set the IO pins. They have the sensor and button for input and the high power relay and low power relay as outputs.
the simple and yet quite explanatory drawing of the circuit, tell us how this stuff will be programmed and how it will run.